More accuracy in display at near 1-to-1 zoom levels. Before this change, when you toggled between a zoom level of 3x and 4x, the horizontal offset calculation seemed slightly biased to the left (say) then when you toggled between 4x and 6x, the bias was to the right. In other words, the display should zoom in on exactly the middle of the visible waveform, but before this change, it didn't.

+++ src/sound.h
@@ -36,5 +36,5 @@ public:
 
     unsigned GetLength();
 
-    void CalcDisplayData(int start_sample_idx, int16_t *mins, int16_t *maxes, unsigned width_in_pixels, unsigned samples_per_pixel);
+    void CalcDisplayData(int start_sample_idx, int16_t *mins, int16_t *maxes, unsigned width_in_pixels, double samples_per_pixel);
 };
